Text To speech functionality
Integrated text to speech functionality into voicify app.
This feature can be toggled via the TTS master switch on the main activity page of the application.

Currently text-to-speech will speak when tapping the "Listen" button that activates the listener.

Known issue: TTS object instance is unable to run in background when closing the app, leading to "tts service not bound" error