Text to Speech integration as a service
Created text to speech service that runs in the background when master TTS toggle is active in voicify app.

To call service from within MainActivity, use speakerTask() with a string as the input parameter.

This functionality was implemented within the openApp method to provide audible feedback for voice commands such as "Opening Chrome".