Update AndroidManifest.xml - Crash fix
Modified AndroidManifest to include an additional field in the "VoiceToActionService" which resolves an issue encountered where voicify app would crash and in the Accessibility settings on the device, would display "Not working. Tap for info". This would also prevent the service from starting sometimes unless a user  activates and deactivates the service twice.

The shortcut button on the navigation bar can now start Voicify correctly every time