Moveable switch bar bug fix
Fixed issue with movable switch which prevented a user from touching and interacting with on-screen elements under the overlay view which would essentially render the entire screen unusable.

The switch bar can now be dragged anywhere on the screen, the switch buttons are clickable and the underlying screen functionality now works together.