Action button styling and spacing
Updated action buttons design (start, record, play) to be more visually appealing and spaced to minimize the chance of a user clicking the incorrect button when interacting with the screen.

A drawable file was created to add some rounded corners to the buttons on the overlay, however, it does not seem to apply properly on the overlay service despite being implemented exactly the same way as the tooltip labels. This should be looked into further if there is time.